requirements.txt
Flask==3.0.2
gspread==6.0.0
oauth2client==4.1.3
python-dotenv==1.0.1
gunicorn==21.2.0